removed old output checker and updated class-based output checker with Gwyn's new code
